Task 1: Within the Main.java file, add a line of code within the main method that will output the text "Hello World!" to the screen. Compile and execute the program to check that it works as required.
Task 2 - Now amend Main.java to enable the program to enter a user's name and store in a variable.
Change the output of "Hello World!" so that the program outputs a personal greeting to the name
entered (for example "Hello Simon!")

Task 3 - Design a program that asks the user to enter values for the length and height of a rectangle and
then outputs the perimeter and area of the rectangle to the screen

Task 4 - Write a program that calculates and displays the average of two numbers entered. Check this
program works for whole and decimal values

Task 5 - Write a program that allows the user to enter two integer values. The program should then output the original order they were entered, then output the values in reverse order (swapped)