Task 1: Within the Main.java file, add a line of code within the main method that will output the text "Hello World!" to the screen. Compile and execute the program to check that it works as required.
Task 2 - Now amend Main.java to enable the program to enter a user's name and store in a variable.
Change the output of "Hello World!" so that the program outputs a personal greeting to the name
entered (for example "Hello Simon!")

Task 3 - Design a program that asks the user to enter values for the length and height of a rectangle and
then outputs the perimeter and area of the rectangle to the screen

Task 4 - Write a program that calculates and displays the average of two numbers entered. Check this
program works for whole and decimal values

Task 5 - Write a program that allows the user to enter two integer values. The program should then output the original order they were entered, then output the values in reverse order (swapped)

Task 6 - Design a program that allows a user to type in their age. The program should then work out which
year they were born and output this to the screen.

Task 7 - Now modify your solution to question 6, and instead ask the user to type in their date of birth. The
program should then calculate how many days old they are (from today's date). Consider how you
will store the components of the users date of birth (DOB) - which consists of day, month and year.

Task 8 - Write a program that will convert the number of feet entered to miles. Note: there are 5280 feet to
1 mile.

Task 9 - Write a program that will calculate how much interest will be paid on money invested in a savings
account. The fixed interest rate is 1%. The user should enter the deposit amount and the program
should display how much money the user will have after the interest is paid at the end of the year.